Compile with cairo_jpg together
